patch: move FactoryConfig struct to it's own module
